@@ -38,7 +38,7 @@
     "markdown-it": "^8.3.1",
     "markdown-it-katex": "^2.0.3",
     "moment": "^2.18.1",
-    "next": "^4.0",
+    "next": "^5.0",
     "prop-types": "^15.6.0",
     "react": "^16.2",
     "react-dom": "^16.2",
