@@ -428,4 +428,4 @@ Button.extendButtonStyle = (buttonStyle, customStyle) => {
   return newStyle;
 };
 
-module.exports = Button;
+export default Button;
