@@ -9,11 +9,11 @@
  * before we get a better button experience.
  */
 
-const React = require("react");
-const { StyleSheet, css } = require("aphrodite");
+import React from "react";
+import { StyleSheet, css } from "aphrodite";
 import PropTypes from "prop-types";
 
-const globalStyles = require("../../styles.js");
+import globalStyles from "../../styles.js";
 const { colors, wbColors, wbTypography } = globalStyles;
 
 const childrenType = PropTypes.oneOfType([
@@ -433,4 +433,4 @@ Button.extendButtonStyle = (buttonStyle, customStyle) => {
   return newStyle;
 };
 
-module.exports = Button;
+export default Button;
