@@ -538,7 +538,7 @@ const editorStyles = StyleSheet.create({
   },
 });
 
-module.exports = {
+export default {
   borderRadius,
   buttons,
   colors,
